Fix style, get new body source method up and working.
