+++ src/serac/physics/tests/test_functional_weak_form.cpp
@@ -5,76 +5,46 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 
 #include <cstddef>
-#include <cmath>
 #include <memory>
+#include <iostream>
 #include <string>
 #include <vector>
 
 #include "gtest/gtest.h"
 #include "mpi.h"
-
 #include "mfem.hpp"
+
 #include "serac/infrastructure/application_manager.hpp"
-#include "serac/physics/materials/solid_material.hpp"
 #include "serac/physics/mesh.hpp"
 #include "serac/physics/state/state_manager.hpp"
 #include "serac/physics/tests/physics_test_utils.hpp"
-#include "serac/physics/solid_residual.hpp"
-#include "serac/infrastructure/accelerator.hpp"
-#include "serac/numerics/functional/domain.hpp"
+#include "serac/physics/functional_weak_form.hpp"
 #include "serac/numerics/functional/finite_element.hpp"  // for H1
+#include "serac/numerics/functional/functional.hpp"
 #include "serac/numerics/functional/tensor.hpp"
+#include "serac/numerics/functional/tuple.hpp"
 #include "serac/physics/common.hpp"
 #include "serac/physics/field_types.hpp"
-#include "serac/physics/residual.hpp"
 #include "serac/physics/state/finite_element_dual.hpp"
 #include "serac/physics/state/finite_element_state.hpp"
 
 auto element_shape = mfem::Element::QUADRILATERAL;
 
-namespace serac {
-
-struct NeoHookeanWithFieldWithRateForTesting {
-  using State = Empty;  ///< this material has no internal variables
-
-  template <typename T1, typename T2, int dim>
-  SERAC_HOST_DEVICE auto pkStress(double /*dt*/, State& /* state */, const tensor<T1, dim, dim>& du_dX,
-                                  const tensor<T2, dim, dim>& /*dv_dX*/) const
-  {
-    using std::log1p;
-    constexpr auto I = Identity<dim>();
-    auto lambda = K - (2.0 / 3.0) * G;
-    auto B_minus_I = dot(du_dX, transpose(du_dX)) + transpose(du_dX) + du_dX;
-
-    auto logJ = log1p(detApIm1(du_dX));
-    // Kirchoff stress, in form that avoids cancellation error when F is near I
-    auto TK = lambda * logJ * I + G * B_minus_I;
-
-    // Pull back to Piola
-    auto F = du_dX + I;
-    return dot(TK, inv(transpose(F)));
-  }
-
-  SERAC_HOST_DEVICE auto density() const { return Rho; }
-
-  double K;    ///< bulk modulus
-  double G;    ///< shear modulus
-  double Rho;  ///< density
-};
-
-}  // namespace serac
-
-struct ResidualFixture : public testing::Test {
+struct WeakFormFixture : public testing::Test {
   static constexpr int dim = 2;
   static constexpr int disp_order = 1;
 
   using VectorSpace = serac::H1<disp_order, dim>;
   using DensitySpace = serac::L2<disp_order - 1>;
 
-  using SolidMaterial = serac::solid_mechanics::NeoHookeanWithFieldDensity;
-  using SolidRateMaterial = serac::NeoHookeanWithFieldWithRateForTesting;
+  enum STATE
+  {
+    DISP,
+    VELO,
+    NUM_STATES
+  };
 
-  enum PARAMS
+  enum PAR
   {
     DENSITY
   };
@@ -91,13 +61,12 @@ struct ResidualFixture : public testing::Test {
 
     serac::FiniteElementState disp = serac::StateManager::newState(VectorSpace{}, "displacement", mesh->tag());
     serac::FiniteElementState velo = serac::StateManager::newState(VectorSpace{}, "velocity", mesh->tag());
-    serac::FiniteElementState accel = serac::StateManager::newState(VectorSpace{}, "acceleration", mesh->tag());
     serac::FiniteElementState density = serac::StateManager::newState(DensitySpace{}, "density", mesh->tag());
 
     shape_disp = std::make_unique<serac::FiniteElementState>(mesh->newShapeDisplacement());
     shape_disp_dual = std::make_unique<serac::FiniteElementDual>(mesh->newShapeDisplacementDual());
 
-    states = {disp, velo, accel};
+    states = {disp, velo};
     params = {density};
 
     for (auto s : states) {
@@ -110,28 +79,30 @@ struct ResidualFixture : public testing::Test {
     state_tangents = states;
     param_tangents = params;
 
-    std::string physics_name = "solid";
+    std::string physics_name = "fake_physics";
 
-    auto solid_mechanics_residual = std::make_shared<SolidResidualT>(
-        physics_name, mesh, states[SolidResidualT::DISPLACEMENT].space(), getSpaces(params));
-    SolidMaterial mat;
-    mat.K = 1.0;
-    mat.G = 0.5;
-    SolidRateMaterial rate_mat;
-    rate_mat.K = 1.0;
-    rate_mat.G = 0.5;
-    rate_mat.Rho = 1.5;
+    using TrialSpace = VectorSpace;
 
-    solid_mechanics_residual->setMaterial(serac::DependsOn<0>{}, mesh->entireBodyName(), mat);
-    solid_mechanics_residual->setRateMaterial(serac::DependsOn<>{}, mesh->entireBodyName(), rate_mat);
+    using WeakFormT =
+        serac::FunctionalWeakForm<dim, TrialSpace, serac::Parameters<VectorSpace, VectorSpace, DensitySpace>>;
 
-    // apply traction boundary conditions
+    std::vector<const mfem::ParFiniteElementSpace*> inputs{&states[STATE::DISP].space(), &states[STATE::VELO].space(),
+                                                           &params[PAR::DENSITY].space()};
+
+    auto f_weak_form = std::make_shared<WeakFormT>(physics_name, mesh, states[STATE::DISP].space(), inputs);
+
+    // apply some traction boundary conditions
+
     std::string surface_name = "side";
     mesh->addDomainOfBoundaryElements(surface_name, serac::by_attr<dim>(1));
-    solid_mechanics_residual->addBoundaryIntegral(surface_name, [](auto /*t*/, auto /*x*/, auto n) { return 1.0 * n; });
-    solid_mechanics_residual->addPressure(surface_name, [](auto /*t*/, auto /*x*/) { return 0.6; });
+
+    f_weak_form->addBoundaryIntegral(surface_name, [](double /*t*/, auto /*x*/, auto n) { return 1.0 * n; });
+    f_weak_form->addBodySource(serac::DependsOn<0>{}, mesh->entireBodyName(),
+                               [](double /*t*/, auto /*x*/, auto u) { return u; });
+    f_weak_form->addBodySource(mesh->entireBodyName(), [](double /*t*/, auto x) { return 0.5 * x; });
 
     // initialize fields for testing
+
     for (auto& s : state_tangents) {
       pseudoRand(s);
     }
@@ -139,27 +110,25 @@ struct ResidualFixture : public testing::Test {
       pseudoRand(p);
     }
 
-    state_duals[0] = 1.0;  // used to test that vjp acts via +=, add initial value to shape displacement dual
+    state_duals[DISP] = 1.0;
+    state_duals[VELO] = 2.0;
+    param_duals[DENSITY] = 3.0;
 
-    states[SolidResidualT::DISPLACEMENT].setFromFieldFunction([](serac::tensor<double, dim> x) {
+    states[DISP].setFromFieldFunction([](serac::tensor<double, dim> x) {
       auto u = 0.1 * x;
       return u;
     });
-    states[SolidResidualT::VELOCITY].setFromFieldFunction([](serac::tensor<double, dim> x) {
-      auto u = -0.1 * x;
-      return u;
-    });
-    states[SolidResidualT::ACCELERATION].setFromFieldFunction([](serac::tensor<double, dim> x) {
+
+    states[VELO].setFromFieldFunction([](serac::tensor<double, dim> x) {
       auto u = -0.01 * x;
       return u;
     });
-    params[0] = 1.2;
+
+    params[DENSITY] = 1.2;
 
     // residual is abstract Residual class to ensure usage only through BasePhysics interface
-    residual = solid_mechanics_residual;
+    weak_form = f_weak_form;
   }
-
-  using SolidResidualT = serac::SolidResidual<disp_order, dim, serac::Parameters<DensitySpace>>;
 
   const double time = 0.0;
   const double dt = 1.0;
@@ -168,7 +137,7 @@ struct ResidualFixture : public testing::Test {
 
   axom::sidre::DataStore datastore;
   std::shared_ptr<serac::Mesh> mesh;
-  std::shared_ptr<serac::Residual> residual;
+  std::shared_ptr<serac::WeakForm> weak_form;
 
   std::unique_ptr<serac::FiniteElementState> shape_disp;
   std::unique_ptr<serac::FiniteElementDual> shape_disp_dual;
@@ -183,13 +152,14 @@ struct ResidualFixture : public testing::Test {
   std::vector<serac::FiniteElementState> param_tangents;
 };
 
-TEST_F(ResidualFixture, VjpConsistency)
+TEST_F(WeakFormFixture, VjpConsistency)
 {
   // initialize the displacement and acceleration to a non-trivial field
   auto input_fields = getConstFieldPointers(states, params);
 
-  serac::FiniteElementDual res_vector(states[SolidResidualT::DISPLACEMENT].space(), "residual");
-  res_vector = residual->residual(time, dt, shape_disp.get(), input_fields);
+  serac::FiniteElementDual res_vector(states[DISP].space(), "residual");
+
+  res_vector = weak_form->residual(time, dt, shape_disp.get(), input_fields);
   ASSERT_NE(0.0, res_vector.Norml2());
 
   auto jacobian_weights = [&](size_t i) {
@@ -203,26 +173,32 @@ TEST_F(ResidualFixture, VjpConsistency)
   pseudoRand(v);
   auto field_vjps = getFieldPointers(state_duals, param_duals);
 
-  residual->vjp(time, dt, shape_disp.get(), input_fields, {}, getConstFieldPointers(v), shape_disp_dual.get(),
-                field_vjps, {});
+  std::vector<serac::FiniteElementDual> field_vjps_slow;
+  for (auto& vjp : field_vjps) {
+    field_vjps_slow.push_back(*vjp);
+  }
 
   for (size_t i = 0; i < input_fields.size(); ++i) {
-    serac::FiniteElementState vjp_slow = *input_fields[i];
-    vjp_slow = 0.0;
-    auto J = residual->jacobian(time, dt, shape_disp.get(), input_fields, jacobian_weights(i));
-    J->MultTranspose(v, vjp_slow);
-    if (i == 0) vjp_slow += 1.0;  // make sure vjp uses +=
-    EXPECT_NEAR(vjp_slow.Norml2(), field_vjps[i]->Norml2(), 1e-12);
+    serac::FiniteElementDual& vjp = field_vjps_slow[i];
+    auto J = weak_form->jacobian(time, dt, shape_disp.get(), input_fields, jacobian_weights(i));
+    J->AddMultTranspose(v, vjp);
+  }
+  weak_form->vjp(time, dt, shape_disp.get(), input_fields, {}, getConstFieldPointers(v), shape_disp_dual.get(),
+                 field_vjps, {});
+
+  for (size_t i = 0; i < input_fields.size(); ++i) {
+    EXPECT_NEAR(field_vjps_slow[i].Norml2(), field_vjps[i]->Norml2(), 1e-12)
+        << " " << field_vjps_slow[i].name() << std::endl;
   }
 }
 
-TEST_F(ResidualFixture, JvpConsistency)
+TEST_F(WeakFormFixture, JvpConsistency)
 {
   // initialize the displacement and acceleration to a non-trivial field
   auto input_fields = getConstFieldPointers(states, params);
 
-  serac::FiniteElementDual res_vector(states[SolidResidualT::DISPLACEMENT].space(), "residual");
-  res_vector = residual->residual(time, dt, shape_disp.get(), input_fields);
+  serac::FiniteElementDual res_vector(states[DISP].space(), "residual");
+  res_vector = weak_form->residual(time, dt, shape_disp.get(), input_fields);
   ASSERT_NE(0.0, res_vector.Norml2());
 
   auto jacobianWeights = [&](size_t i) {
@@ -241,34 +217,32 @@ TEST_F(ResidualFixture, JvpConsistency)
     return field_tangents;
   };
 
-  serac::FiniteElementDual jvp_slow(states[SolidResidualT::DISPLACEMENT].space(), "jvp_slow");
-  serac::FiniteElementDual jvp(states[SolidResidualT::DISPLACEMENT].space(), "jvp");
+  serac::FiniteElementDual jvp_slow(states[DISP].space(), "jvp_slow");
+  serac::FiniteElementDual jvp(states[DISP].space(), "jvp");
   jvp = 4.0;  // set to some value to test jvp resets these values
   auto jvps = getFieldPointers(jvp);
 
   auto field_tangents = getConstFieldPointers(state_tangents, param_tangents);
 
   for (size_t i = 0; i < input_fields.size(); ++i) {
-    auto J = residual->jacobian(time, dt, shape_disp.get(), input_fields, jacobianWeights(i));
+    auto J = weak_form->jacobian(time, dt, shape_disp.get(), input_fields, jacobianWeights(i));
     J->Mult(*field_tangents[i], jvp_slow);
-    residual->jvp(time, dt, shape_disp.get(), input_fields, {}, nullptr, selectStates(i), {}, jvps);
+    weak_form->jvp(time, dt, shape_disp.get(), input_fields, {}, nullptr, selectStates(i), {}, jvps);
     EXPECT_NEAR(jvp_slow.Norml2(), jvp.Norml2(), 1e-12);
   }
 
   // test jacobians in weighted combinations
   {
-    field_tangents[SolidResidualT::VELOCITY] = nullptr;
-    field_tangents[size_t(SolidResidualT::NUM_STATES) + size_t(DENSITY)] = nullptr;
+    field_tangents[NUM_STATES] = nullptr;
 
-    double acceleration_factor = 0.2;
-    std::vector<double> jacobian_weights = {1.0, 0.0, acceleration_factor, 0.0};
+    double velo_factor = 0.2;
+    std::vector<double> jacobian_weights = {1.0, velo_factor, 0.0};
 
-    auto J = residual->jacobian(time, dt, shape_disp.get(), input_fields, jacobian_weights);
-    J->Mult(*field_tangents[SolidResidualT::DISPLACEMENT], jvp_slow);
+    auto J = weak_form->jacobian(time, dt, shape_disp.get(), input_fields, jacobian_weights);
+    J->Mult(*field_tangents[DISP], jvp_slow);
 
-    state_tangents[SolidResidualT::ACCELERATION] *= acceleration_factor;
-
-    residual->jvp(time, dt, shape_disp.get(), input_fields, {}, nullptr, field_tangents, {}, jvps);
+    state_tangents[VELO] *= velo_factor;
+    weak_form->jvp(time, dt, shape_disp.get(), input_fields, {}, nullptr, field_tangents, {}, jvps);
     EXPECT_NEAR(jvp_slow.Norml2(), jvp.Norml2(), 1e-12);
   }
 }
